requirements.txt
asgiref
crispy-bootstrap5  
dj-database-url    
Django
django-crispy-forms
django-email       
django-environ     
django-filter      
djangorestframework
EasyProcess        
entrypoint2        
Markdown
MouseInfo
mss
mysql-connector-python
numpy
opencv-contrib-python
Pillow
protobuf
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
pyscreenshot
PyScreeze
pytweening
pytz
sqlparse
typing_extensions
tzdata